# optimize struct field encoding